requirements.txt
ament_copyright==0.12.11
ament_flake8==0.12.11
ament_index_python==1.4.0
ament_pep257==0.12.11
apriltag_msgs==2.0.1
cv_bridge==3.2.1
geometry_msgs==4.2.4
launch_ros==0.19.7
numpy==1.21.5
opencv_python==4.10.0.84
opencv_python_headless==4.8.1.78
pyserial==3.5
pytest==6.2.5
rclpy==3.3.14
sensor_msgs==4.2.4
setuptools==59.6.0
std_msgs==4.2.4
tf2_msgs==0.25.8
tf2_ros_py==0.25.8
tf_transformations==1.1.0
